# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUserCredentialResponse": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					},
					"expired": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"access_token",
					"refresh_token",
					"expired"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBaseErrorDetail": {
				"properties": {
					"property": {
						"type": "string"
					},
					"messsage": {
						"type": "string"
					}
				},
				"required": [
					"property",
					"messsage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBaseErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/IBaseErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserCredential": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TUserRole": {
				"type": "string",
				"enum": [
					"admin",
					"user"
				]
			},
			"IUserAccount": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"display_name": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/TUserRole"
					}
				},
				"required": [
					"id",
					"username",
					"display_name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBaseEmptyResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "Demo Express Typescript",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserCredentialResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBaseErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Please try again with the correct username and password."
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCredential"
							}
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserAccount"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBaseErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Request had invalid authentication credentials."
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBaseEmptyResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBaseErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Your request token has expired. Please obtain a new one and try again."
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "refresh_token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3100/api",
			"description": "Local development"
		}
	]
}